Mobile: Login in tab, hide tab bar when signed in, rename tab to Login; Auth context; Backend: docker env_file
